Add comprehensive XML docs and improve code clarity
Extensive XML documentation was added to all public interfaces, classes, methods, and properties across the Dataverse client library, greatly enhancing IntelliSense and API discoverability. The DataverseQueryBuilder, ExpandBuilder, and FilterBuilder classes are now fully documented, with usage examples and remarks on thread safety. Naming inconsistencies (e.g., InludeAnnotations → IncludeAnnotations) were corrected. Service registration and core types now have clear summaries and parameter descriptions. These changes do not affect runtime behavior but significantly improve maintainability and developer experience.
