+++ src/Dataverse/AuthenticationTokenProvider/ManagedIdentityTokenProvider.cs
@@ -4,6 +4,12 @@ using Microsoft.Extensions.Caching.Memory;
 
 namespace Mavrix.Common.Dataverse.AuthenticationTokenProvider
 {
+	/// <summary>
+	/// Provides Dataverse access tokens using Azure Managed Identity with in-memory caching.
+	/// </summary>
+	/// <remarks>
+	/// Tokens are cached per scope until one minute before expiration. Instances are safe for concurrent calls; disposal is expected to be handled by the host at shutdown.
+	/// </remarks>
 	public class ManagedIdentityTokenProvider : IAzureTokenProvider, IDisposable
 	{
 		private readonly TokenCredential _credential;
@@ -11,17 +17,26 @@ namespace Mavrix.Common.Dataverse.AuthenticationTokenProvider
 
 		private readonly SemaphoreSlim SemaphoreSlim = new(1, 1);
 
+		/// <summary>
+		/// Initializes a new instance of <see cref="ManagedIdentityTokenProvider"/> using the default Azure credential chain.
+		/// </summary>
+		/// <param name="memoryCache">Cache for storing tokens per scope.</param>
 		public ManagedIdentityTokenProvider(IMemoryCache memoryCache)
 		{
 			_credential = new DefaultAzureCredential();
 			_memoryCache = memoryCache;
 		}
 
+		/// <summary>
+		/// Gets the underlying <see cref="TokenCredential"/> used for token acquisition.
+		/// </summary>
 		public TokenCredential GetTokenCredential() => _credential;
 
+		/// <inheritdoc />
 		public async ValueTask<string> GetTokenAsync(string scope, CancellationToken cancellationToken) =>
 			(await GetAccessTokenAsync(scope, cancellationToken)).Token;
 
+		/// <inheritdoc />
 		public async ValueTask<AccessToken> GetAccessTokenAsync(string scope, CancellationToken cancellationToken)
 		{
 			if (_memoryCache.TryGetValue(scope, out AccessToken accessToken))
@@ -47,6 +62,9 @@ namespace Mavrix.Common.Dataverse.AuthenticationTokenProvider
 			}
 		}
 
+		/// <summary>
+		/// Releases resources used by the token provider.
+		/// </summary>
 		public void Dispose()
 		{
 			GC.SuppressFinalize(this);
