+++ src/Dataverse/QueryBuilder/ExpandBuilder.cs
@@ -2,31 +2,69 @@
 
 namespace Mavrix.Common.Dataverse.QueryBuilder
 {
+	/// <summary>
+	/// Builds <c>$expand</c> clauses for Dataverse OData queries, including nested expands.
+	/// </summary>
+	/// <remarks>
+	/// Inputs should be valid OData fragments and pre-encoded as needed; escaping is not performed.
+	/// Instances are not thread-safe and are intended for one-time use per query.
+	/// </remarks>
 	public class ExpandBuilder(string property)
 	{
+		/// <summary>
+		/// Gets the primary navigation property to expand.
+		/// </summary>
 		private string Property { get; set; } = property;
+		/// <summary>
+		/// Gets the properties to select within the expanded entity.
+		/// </summary>
 		private string[] Select { get; set; } = [];
+		/// <summary>
+		/// Gets the filter applied within the expanded entity.
+		/// </summary>
 		private string? Filter { get; set; }
+		/// <summary>
+		/// Gets nested expand definitions.
+		/// </summary>
 		private List<ExpandBuilder> NestedExpands { get; set; } = [];
 
+		/// <summary>
+		/// Specifies the properties to select in the expanded entity.
+		/// </summary>
+		/// <param name="select">Property names to include.</param>
+		/// <returns>The current builder instance.</returns>
 		public ExpandBuilder WithSelect(params string[] select)
 		{
 			Select = select;
 			return this;
 		}
 
+		/// <summary>
+		/// Applies a filter within the expanded entity.
+		/// </summary>
+		/// <param name="filter">The OData filter expression.</param>
+		/// <returns>The current builder instance.</returns>
 		public ExpandBuilder WithFilter(string filter)
 		{
 			Filter = filter;
 			return this;
 		}
 
+		/// <summary>
+		/// Adds a nested expand to include related entities of the expanded entity.
+		/// </summary>
+		/// <param name="nestedExpand">The nested expand builder.</param>
+		/// <returns>The current builder instance.</returns>
 		public ExpandBuilder AddNestedExpand(ExpandBuilder nestedExpand)
 		{
 			NestedExpands.Add(nestedExpand);
 			return this;
 		}
 
+		/// <summary>
+		/// Builds the <c>$expand</c> expression for this builder and any nested builders.
+		/// </summary>
+		/// <returns>The composed expand expression.</returns>
 		public string Build()
 		{
 			var expandExpression = new StringBuilder(Property);
@@ -66,6 +104,10 @@ namespace Mavrix.Common.Dataverse.QueryBuilder
 			return expandExpression.ToString();
 		}
 
+		/// <summary>
+		/// Returns the built <c>$expand</c> expression.
+		/// </summary>
+		/// <returns>The composed expand expression.</returns>
 		public override string ToString() => Build();
 	}
 }
