+++ src/Dataverse/ServiceCollectionExtensions.cs
@@ -14,8 +14,19 @@ using System.Text.Json;
 
 namespace Mavrix.Common.Dataverse
 {
+	/// <summary>
+	/// Extension methods for registering Dataverse clients, repositories, and related services.
+	/// </summary>
 	public static class ServiceCollectionExtensions
 	{
+		/// <summary>
+		/// Registers the Dataverse HTTP client, token provider, serializer options, and retry policy.
+		/// </summary>
+		/// <param name="services">The service collection to add registrations to.</param>
+		/// <param name="configuration">Configuration source containing Dataverse settings.</param>
+		/// <param name="configureSerializer">Optional callback to customize serializer options.</param>
+		/// <param name="useManagedIdentity">Whether to register the managed identity token provider.</param>
+		/// <returns>The service collection for chaining.</returns>
 		public static IServiceCollection AddDataverseClient(
 			this IServiceCollection services, 
 			ConfigurationManager configuration, 
@@ -44,6 +55,11 @@ namespace Mavrix.Common.Dataverse
 			return services;
 		}
 
+		/// <summary>
+		/// Adds default logging filters for Dataverse HTTP client noise reduction.
+		/// </summary>
+		/// <param name="builder">The logging builder to configure.</param>
+		/// <returns>The logging builder for chaining.</returns>
 		public static ILoggingBuilder AddDataverseDefaultLoggingSettings(this ILoggingBuilder builder)
 		{
 			builder.AddFilter("System.Net.Http.HttpClient.IDataverseHttpClient.LogicalHandler", LogLevel.Warning);
@@ -51,6 +67,11 @@ namespace Mavrix.Common.Dataverse
 			return builder;
 		}
 
+		/// <summary>
+		/// Adds a retry handler for HTTP 429 (Too Many Requests) responses to the Dataverse client.
+		/// </summary>
+		/// <param name="builder">The HTTP client builder.</param>
+		/// <returns>The resilience pipeline builder for further configuration.</returns>
 		public static IHttpResiliencePipelineBuilder AddTooManyRequestRetryHandler(this IHttpClientBuilder builder)
 		{
 			return builder.AddResilienceHandler("default", static configure =>
@@ -67,6 +88,12 @@ namespace Mavrix.Common.Dataverse
 			});
 		}
 
+		/// <summary>
+		/// Registers a Dataverse repository for the specified entity type.
+		/// </summary>
+		/// <typeparam name="T">The DTO type representing the Dataverse set.</typeparam>
+		/// <param name="services">The service collection to add registrations to.</param>
+		/// <returns>The service collection for chaining.</returns>
 		public static IServiceCollection AddDataverseRepository<T>(this IServiceCollection services) where T : DataverseTable
 		{
 			services.TryAdd(new ServiceDescriptor(typeof(IDataverseRepository<T>), sp =>
@@ -78,6 +105,12 @@ namespace Mavrix.Common.Dataverse
 			return services;
 		}
 
+		/// <summary>
+		/// Registers a JSON serializer options configurator for Dataverse serialization.
+		/// </summary>
+		/// <param name="services">The service collection to add registrations to.</param>
+		/// <param name="configurator">Configurator instance to apply.</param>
+		/// <returns>The service collection for chaining.</returns>
 		public static IServiceCollection AddDataverseJsonSerializerConfigurator(this IServiceCollection services, IDataverseJsonSerializerOptionsConfigurator configurator)
 		{
 			services.AddSingleton(configurator);
